src/main/kotlin/ch/kleis/lcaplugin/core/lang/evaluator/step/Reduce.kt
package ch.kleis.lcaplugin.core.lang.evaluator.step

import ch.kleis.lcaplugin.core.lang.SymbolTable
import ch.kleis.lcaplugin.core.lang.evaluator.EvaluatorException
import ch.kleis.lcaplugin.core.lang.evaluator.Helper
import ch.kleis.lcaplugin.core.lang.evaluator.reducer.LcaExpressionReducer
import ch.kleis.lcaplugin.core.lang.evaluator.reducer.TemplateExpressionReducer
import ch.kleis.lcaplugin.core.lang.expression.*

class Reduce(
    symbolTable: SymbolTable,
) {
    private val lcaReducer = LcaExpressionReducer(
        symbolTable.data,
    )
    private val templateReducer = TemplateExpressionReducer(
        symbolTable.data,
    )

    fun apply(expression: EProcessTemplateApplication): EProcessFinal {
        val reduced = templateReducer.reduceTemplateApplication(expression)
        val unboundedReferences = Helper().allRequiredRefs(reduced)
        if (unboundedReferences.isNotEmpty()) {
            throw EvaluatorException("unbounded references: $unboundedReferences")
        }
        return reduced
    }

    fun apply(expression: ESubstanceCharacterization): ESubstanceCharacterization {
        val reduced = lcaReducer.reduceSubstanceCharacterization(expression)
        val unboundedReferences = Helper().allRequiredRefs(reduced)
        if (unboundedReferences.isNotEmpty()) {
            throw EvaluatorException("unbounded references: $unboundedReferences")
        }
        return reduced
    }


}
